# [Enhancement] optimize aggregate function window_funnel