--- a/be/src/runtime/memory/memory_resource.h
+++ b/be/src/runtime/memory/memory_resource.h
@@ -26,6 +26,10 @@ public:
               _upstream(upstream) {
         ASAN_POISON_MEMORY_REGION(_stack_addr_start, sz);
     }
+    virtual ~stack_memory_resource() {
+        size_t sz = (uint8_t*)_stack_addr_end - (uint8_t*)_stack_addr_start;
+        ASAN_UNPOISON_MEMORY_REGION(_stack_addr_start, sz);
+    }
 
 private:
     void* _stack_addr_start;