## What type of PR is this：
- [ ] bug
- [ ] feature
- [x] enhancement
- [ ] refactor
- [ ] others

## Which issues of this PR fixes ：
Fixes #

## Problem Summary(Required) ：
1. reduce memory alloc by PMR (alloc by stack)
2. reduce some function call (ArrayColumn::offsets/elements)
3. reduce some virtual function call (eg: append(DatumArray))


case:
```
mysql> select level, count(*) uv from  (     select lo_orderkey, window_funnel(3600000, lo_orderdate, 0, [lo_linenumber=1, lo_linenumber=2]) as level      from lineorder      group by lo_orderkey ) AS x  group by level;
```
16 core for each BE

|                                  | baseline (sec) | patched (sec) |
| -------------------------------- | -------------- | ------------- |
| colocate one stage AGG (1FE 3BE) | 9.36           | 5.33          |
| one stage AGG (1FE 4BE)          | 8.77           | 5.74          |
| two stage AGG (1FE 4BE)          | 14.65          | 7.11          |