@wanpengfei-git 2022-07-28
run starrocks_admit_test